updates to PR comments
